% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mgf.R
\name{parse_mgf}
\alias{parse_mgf}
\title{Parse mgf}
\usage{
parse_mgf(mgf_path)
}
\arguments{
\item{mgf_path}{path to mgf}
}
\value{
data.table with cols pepmass, charge, scan,
    and col containing json representation of mass/intensity
}
\description{
Parse mgf
}
